move mod for generated code
