+++ src/network_filter_list.rs
@@ -4,9 +4,9 @@ use std::{collections::HashMap, collections::HashSet, fmt};
 
 use flatbuffers::ForwardsUOffset;
 
-use crate::filters::fb_network::flat::fb;
 use crate::filters::fb_network::FlatNetworkFilter;
 use crate::filters::filter_data_context::FilterDataContext;
+use crate::filters::flatbuffer_generated::fb;
 use crate::filters::network::{
     NetworkFilter, NetworkFilterMask, NetworkFilterMaskHelper, NetworkMatchable,
 };
